[chore] Rename folder 'views' to 'screens'
