+++ rental_app/lib/main.dart
@@ -1,7 +1,7 @@
-import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:rental_app/views/authentication_checker.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:rental_app/screens/authentication_checker.dart';
 
 import 'config/theme.dart';
