+++ rental_app/lib/screens/authentication_checker.dart
