+++ rental_app/lib/screens/error_screen.dart
