+++ rental_app/lib/screens/login_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:rental_app/providers/authentication_provider.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:rental_app/views/reset_password_page.dart';
+import 'package:rental_app/providers/authentication_provider.dart';
+import 'package:rental_app/screens/reset_password_page.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({Key? key}) : super(key: key);
